backend/requirements.txt
# DocLoop backend — Python dependencies
# Pin major versions; let patch versions float for security fixes.
# Install: pip install -r requirements.txt
# Dev:     pip install -r requirements-dev.txt

# ── Web framework ──
fastapi==0.115.*
uvicorn[standard]==0.30.*          # includes uvloop + httptools for prod speed
python-multipart==0.0.*            # required for FastAPI file uploads

# ── Database ──
sqlalchemy==2.0.*
asyncpg==0.29.*                    # async PostgreSQL driver
alembic==1.13.*                    # migrations

# ── Config ──
pydantic==2.7.*
pydantic-settings==2.3.*

# ── Redis + async job queue ──
redis[hiredis]==5.0.*              # hiredis for faster parsing
arq==0.25.*                        # ARQ async job queue

# ── Scheduling ──
apscheduler==3.10.*

# ── Auth ──
PyJWT==2.8.*                       # Supabase JWT verification
cryptography==42.*                 # PyJWT dependency for RS256 (future-proofing)

# ── Supabase client (storage + admin auth) ──
supabase==2.5.*

# ── Email ──
resend==2.3.*

# ── SMS ──
twilio==9.1.*

# ── Payments ──
stripe==10.4.*

# ── AI ──
openai==1.35.*

# ── Utilities ──
python-jose[cryptography]==3.3.*   # JWT utilities (alternative to PyJWT if needed)
httpx==0.27.*                      # async HTTP client used by Supabase SDK
tenacity==8.3.*                    # retry logic for external service calls
